refactor(checker): split class heritage diagnostics
