+++ crates/tsz-checker/src/state/state_checking/class.rs
@@ -1871,563 +1871,6 @@ impl<'a> CheckerState<'a> {
         }
     }
 
-    /// TS2509: Check that the base constructor return type is an object type or
-    /// intersection of object types with statically known members.
-    ///
-    /// When a class extends another via a heritage clause, the return type of
-    /// the base constructor must be valid. For example, if `Mix(Private, Private2)`
-    /// returns an intersection that reduces to `never`, this is not a valid base type.
-    fn check_base_constructor_return_type(
-        &mut self,
-        class_data: &tsz_parser::parser::node::ClassData,
-    ) {
-        use crate::diagnostics::{diagnostic_codes, diagnostic_messages, format_message};
-        use tsz_scanner::SyntaxKind;
-
-        let Some(ref heritage_clauses) = class_data.heritage_clauses else {
-            return;
-        };
-
-        // A class declaration may only have ONE `extends` clause; subsequent
-        // ones are TS1172 parser errors. Skip the duplicates so we don't try
-        // to resolve their types and emit cascading TS2304 / TS2509.
-        let mut extends_seen = false;
-
-        for &clause_idx in &heritage_clauses.nodes {
-            let Some(clause_node) = self.ctx.arena.get(clause_idx) else {
-                continue;
-            };
-            let Some(heritage) = self.ctx.arena.get_heritage_clause(clause_node) else {
-                continue;
-            };
-            if heritage.token != SyntaxKind::ExtendsKeyword as u16 {
-                continue;
-            }
-            if extends_seen {
-                continue;
-            }
-            extends_seen = true;
-
-            let Some(&type_idx) = heritage.types.nodes.first() else {
-                continue;
-            };
-
-            let (expr_idx, type_arguments) = if let Some(type_node) = self.ctx.arena.get(type_idx)
-                && let Some(expr_type_args) = self.ctx.arena.get_expr_type_args(type_node)
-            {
-                (
-                    expr_type_args.expression,
-                    expr_type_args.type_arguments.as_ref(),
-                )
-            } else {
-                (type_idx, None)
-            };
-
-            // Get the base instance type (constructor return type)
-            let Some(base_type) = self.base_instance_type_from_expression(expr_idx, type_arguments)
-            else {
-                continue;
-            };
-
-            // Skip for any/error/null — these are permissive.
-            // `class extends null` is valid TS (produces a class with no prototype).
-            if base_type == TypeId::ANY || base_type == TypeId::ERROR || base_type == TypeId::NULL {
-                continue;
-            }
-
-            // Skip for null — `class C extends null` is valid in TypeScript.
-            // tsc does not emit TS2509 for null base types; instead, it only
-            // checks for TS17005 (super call in null-extending class).
-            if base_type == TypeId::NULL {
-                continue;
-            }
-
-            // Skip union base types. When a constructor has multiple construct
-            // signatures (e.g., `Array` with `new(): any[]` and `new<T>(): T[]`),
-            // the resolved base instance type can be a union like `any[] | T[]`.
-            // tsc resolves these to the correct specific return type; our resolution
-            // currently produces the union. Since all constituent types in such
-            // unions are valid object types, suppress TS2509 for unions.
-            if crate::query_boundaries::common::is_union_type(self.ctx.types, base_type) {
-                continue;
-            }
-
-            // Check if the base type is a valid base type. Mixin intersections
-            // with incompatible private property origins are invalid bases even
-            // when the structural intersection has not been fully reduced to
-            // `never` yet.
-            let private_property_conflict =
-                self.intersection_has_private_property_conflict(base_type);
-            if !crate::query_boundaries::class::is_valid_base_type(self.ctx.types, base_type)
-                || private_property_conflict
-            {
-                let type_name = if private_property_conflict {
-                    self.format_type(TypeId::NEVER)
-                } else {
-                    self.format_type(base_type)
-                };
-                let message = format_message(
-                    diagnostic_messages::BASE_CONSTRUCTOR_RETURN_TYPE_IS_NOT_AN_OBJECT_TYPE_OR_INTERSECTION_OF_OBJECT_TYP,
-                    &[&type_name],
-                );
-                self.error_at_node(
-                    expr_idx,
-                    &message,
-                    diagnostic_codes::BASE_CONSTRUCTOR_RETURN_TYPE_IS_NOT_AN_OBJECT_TYPE_OR_INTERSECTION_OF_OBJECT_TYP,
-                );
-            }
-
-            break; // Only check the first extends clause
-        }
-    }
-
-    /// TS2797: A mixin class that extends from a type variable containing an
-    /// abstract construct signature must also be declared 'abstract'.
-    ///
-    /// When a non-abstract class extends from a type variable (type parameter)
-    /// whose constraint includes `abstract new (...)`, the class must be abstract.
-    /// This is the mixin pattern: `class C extends baseClass` where `baseClass: T`
-    /// and `T extends abstract new (...args: any) => any`.
-    fn check_mixin_abstract_construct_constraint(
-        &mut self,
-        class_idx: NodeIndex,
-        class_data: &tsz_parser::parser::node::ClassData,
-    ) {
-        use crate::diagnostics::{diagnostic_codes, diagnostic_messages};
-        use tsz_scanner::SyntaxKind;
-
-        let Some(ref heritage_clauses) = class_data.heritage_clauses else {
-            return;
-        };
-
-        // Skip duplicate `extends` clauses on a class (TS1172 parser error).
-        let mut extends_seen = false;
-
-        for &clause_idx in &heritage_clauses.nodes {
-            let Some(clause_node) = self.ctx.arena.get(clause_idx) else {
-                continue;
-            };
-            let Some(heritage) = self.ctx.arena.get_heritage_clause(clause_node) else {
-                continue;
-            };
-            if heritage.token != SyntaxKind::ExtendsKeyword as u16 {
-                continue;
-            }
-            if extends_seen {
-                continue;
-            }
-            extends_seen = true;
-
-            // Get the extends expression
-            let Some(&type_idx) = heritage.types.nodes.first() else {
-                continue;
-            };
-            let Some(type_node) = self.ctx.arena.get(type_idx) else {
-                continue;
-            };
-
-            let expr_idx =
-                if let Some(expr_type_args) = self.ctx.arena.get_expr_type_args(type_node) {
-                    expr_type_args.expression
-                } else {
-                    type_idx
-                };
-
-            // Try get_type_of_node first; if it returns a usable type, use it.
-            // Otherwise, fall back to resolving the parameter's declared type
-            // annotation directly (workaround for name-merging issues where
-            // get_type_of_node returns ANY).
-            let base_type = self.resolve_heritage_expr_declared_type(expr_idx);
-            if base_type == TypeId::ERROR {
-                return;
-            }
-
-            // Check if the base type is a type parameter with a constraint
-            let Some(constraint_type) =
-                class_query::type_parameter_constraint(self.ctx.types, base_type)
-            else {
-                return;
-            };
-
-            // Check if the constraint has abstract construct signatures
-            if self.constraint_has_abstract_construct(constraint_type) {
-                let error_node = if class_data.name.is_some() {
-                    class_data.name
-                } else {
-                    class_idx
-                };
-                self.error_at_node(
-                    error_node,
-                    diagnostic_messages::A_MIXIN_CLASS_THAT_EXTENDS_FROM_A_TYPE_VARIABLE_CONTAINING_AN_ABSTRACT_CONSTRUCT,
-                    diagnostic_codes::A_MIXIN_CLASS_THAT_EXTENDS_FROM_A_TYPE_VARIABLE_CONTAINING_AN_ABSTRACT_CONSTRUCT,
-                );
-            }
-
-            return;
-        }
-    }
-
-    /// TS2545: A mixin class must have a constructor with a single rest parameter
-    /// of type 'any[]'.
-    ///
-    /// When a class extends a type variable (type parameter), the construct
-    /// signatures of the constraint must each have a single rest parameter whose
-    /// type is `any[]` or `readonly any[]`.  If not, emit TS2545.
-    fn check_mixin_constructor_rest_parameter(
-        &mut self,
-        class_idx: NodeIndex,
-        class_data: &tsz_parser::parser::node::ClassData,
-    ) {
-        use crate::diagnostics::{diagnostic_codes, diagnostic_messages};
-        use tsz_scanner::SyntaxKind;
-
-        let Some(ref heritage_clauses) = class_data.heritage_clauses else {
-            return;
-        };
-
-        // Skip duplicate `extends` clauses on a class (TS1172 parser error).
-        let mut extends_seen = false;
-
-        for &clause_idx in &heritage_clauses.nodes {
-            let Some(clause_node) = self.ctx.arena.get(clause_idx) else {
-                continue;
-            };
-            let Some(heritage) = self.ctx.arena.get_heritage_clause(clause_node) else {
-                continue;
-            };
-            if heritage.token != SyntaxKind::ExtendsKeyword as u16 {
-                continue;
-            }
-            if extends_seen {
-                continue;
-            }
-            extends_seen = true;
-
-            let Some(&type_idx) = heritage.types.nodes.first() else {
-                continue;
-            };
-            let Some(type_node) = self.ctx.arena.get(type_idx) else {
-                continue;
-            };
-
-            let expr_idx =
-                if let Some(expr_type_args) = self.ctx.arena.get_expr_type_args(type_node) {
-                    expr_type_args.expression
-                } else {
-                    type_idx
-                };
-
-            let base_type = self.resolve_heritage_expr_declared_type(expr_idx);
-            if base_type == TypeId::ERROR {
-                return;
-            }
-
-            // Only applies when the base type is a type parameter (mixin pattern)
-            let Some(constraint_type) =
-                class_query::type_parameter_constraint(self.ctx.types, base_type)
-            else {
-                return;
-            };
-
-            // Match tsc's gate: TS2545 is emitted from inside an
-            // `if (baseTypes.length) { addLazyDiagnostic(...) }` block in
-            // `checkClassLikeDeclaration`. When the base constructor's return
-            // type is not an object-like type (e.g. `new () => never`,
-            // `new () => void`), `getBaseTypes` produces an empty list and the
-            // whole block — including the mixin-shape check — is skipped, with
-            // TS2509 carrying the diagnostic instead. Mirror that here so the
-            // two diagnostics don't both fire on the same heritage clause.
-            if let Some(base_instance) = self.base_instance_type_from_expression(expr_idx, None)
-                && !crate::query_boundaries::class::is_valid_base_type(
-                    self.ctx.types,
-                    base_instance,
-                )
-            {
-                return;
-            }
-
-            // Evaluate the constraint type (may be a Lazy type alias or Application)
-            let evaluated = self.evaluate_type_for_assignability(constraint_type);
-
-            // Get construct signatures from the evaluated constraint
-            let construct_sigs = self.collect_construct_signatures_from_evaluated(evaluated);
-            if construct_sigs.is_empty() {
-                return;
-            }
-
-            // Surgical fix for #9729: a plain non-generic single signature
-            // with zero params fails tsc's mixin constructor contract. Other
-            // shapes stay on the pre-existing per-sig path; conditional mixin
-            // helper output can currently evaluate to a zero-param signature
-            // even though tsc accepts it.
-            let single_sig_zero_param = construct_sigs.len() == 1
-                && construct_sigs[0].type_params.is_empty()
-                && construct_sigs[0].params.is_empty()
-                && !class_query::contains_conditional_type(self.ctx.types, constraint_type);
-            let has_invalid_sig = single_sig_zero_param
-                || construct_sigs.iter().any(|sig| {
-                    !sig.params.is_empty() && !self.is_valid_mixin_construct_signature(sig)
-                });
-            if has_invalid_sig {
-                self.error_at_node(
-                    class_idx,
-                    diagnostic_messages::A_MIXIN_CLASS_MUST_HAVE_A_CONSTRUCTOR_WITH_A_SINGLE_REST_PARAMETER_OF_TYPE_ANY,
-                    diagnostic_codes::A_MIXIN_CLASS_MUST_HAVE_A_CONSTRUCTOR_WITH_A_SINGLE_REST_PARAMETER_OF_TYPE_ANY,
-                );
-            }
-
-            return;
-        }
-    }
-
-    /// Collect construct signatures from an already-evaluated type.
-    fn collect_construct_signatures_from_evaluated(
-        &self,
-        type_id: TypeId,
-    ) -> Vec<tsz_solver::CallSignature> {
-        if let Some(sigs) = class_query::construct_signatures_for_type(self.ctx.types, type_id)
-            && !sigs.is_empty()
-        {
-            return sigs;
-        }
-
-        // Intersection: collect from all members
-        if let Some(members) = class_query::intersection_members(self.ctx.types, type_id) {
-            let mut all_sigs = Vec::new();
-            for &member in members.iter() {
-                if let Some(sigs) =
-                    class_query::construct_signatures_for_type(self.ctx.types, member)
-                {
-                    all_sigs.extend(sigs);
-                }
-            }
-            return all_sigs;
-        }
-
-        Vec::new()
-    }
-
-    /// Whether a construct signature is shape-compatible with the mixin-base
-    /// contract: exactly one rest parameter (not optional) whose type is `any`,
-    /// `any[]`, or `readonly any[]`.
-    fn is_valid_mixin_construct_signature(&self, sig: &tsz_solver::CallSignature) -> bool {
-        sig.params.len() == 1
-            && sig.params[0].rest
-            && !sig.params[0].optional
-            && self.is_valid_mixin_rest_param_type(sig.params[0].type_id)
-    }
-
-    /// Check if a rest parameter type is valid for a mixin constructor.
-    /// Accepts `any`, `any[]`, or `readonly any[]`.
-    fn is_valid_mixin_rest_param_type(&self, type_id: TypeId) -> bool {
-        // `any` is valid for mixin rest parameters (e.g., `...args: any`)
-        if type_id == TypeId::ANY {
-            return true;
-        }
-        // `any[]` or `readonly any[]`
-        matches!(
-            class_query::array_element_type(self.ctx.types, type_id),
-            Some(elem) if elem == TypeId::ANY
-        )
-    }
-
-    /// Resolve the declared type for a heritage expression identifier.
-    /// First tries `get_type_of_node`. If that returns ANY (which can happen
-    /// due to symbol name merging), falls back to resolving the identifier's
-    /// symbol, finding its parameter declaration, and evaluating the type
-    /// annotation directly.
-    fn resolve_heritage_expr_declared_type(&mut self, expr_idx: NodeIndex) -> TypeId {
-        let base_type = self.get_type_of_node(expr_idx);
-        if base_type != TypeId::ANY {
-            return base_type;
-        }
-
-        // Fallback: resolve via parameter declaration's type annotation
-        let Some(sym_id) = self.resolve_identifier_symbol(expr_idx) else {
-            return TypeId::ANY;
-        };
-        let Some(symbol) = self.ctx.binder.get_symbol(sym_id) else {
-            return TypeId::ANY;
-        };
-        let Some(&decl_idx) = symbol.declarations.first() else {
-            return TypeId::ANY;
-        };
-        let Some(decl_node) = self.ctx.arena.get(decl_idx) else {
-            return TypeId::ANY;
-        };
-        let Some(param_data) = self.ctx.arena.get_parameter(decl_node) else {
-            return TypeId::ANY;
-        };
-        let type_ann = param_data.type_annotation;
-        if type_ann == NodeIndex::NONE {
-            return TypeId::ANY;
-        }
-        self.get_type_from_type_node(type_ann)
-    }
-
-    /// Check if a constraint type (or any member of an intersection constraint)
-    /// contains abstract construct signatures.
-    fn constraint_has_abstract_construct(&self, constraint_type: TypeId) -> bool {
-        // Direct callable check
-        if let Some(callable) =
-            class_query::callable_shape_for_type(self.ctx.types, constraint_type)
-            && callable.is_abstract
-            && !callable.construct_signatures.is_empty()
-        {
-            return true;
-        }
-
-        // Intersection: check each member
-        if let Some(members) = class_query::intersection_members(self.ctx.types, constraint_type) {
-            for &member in members.iter() {
-                if let Some(callable) = class_query::callable_shape_for_type(self.ctx.types, member)
-                    && callable.is_abstract
-                    && !callable.construct_signatures.is_empty()
-                {
-                    return true;
-                }
-            }
-        }
-
-        false
-    }
-
-    /// Check if an anonymous class is exported (via export default modifier or parent export node).
-    fn is_class_exported_default(
-        &self,
-        class_idx: NodeIndex,
-        modifiers: &Option<tsz_parser::parser::NodeList>,
-    ) -> bool {
-        use tsz_scanner::SyntaxKind;
-        // Check for export + default modifiers on the class itself
-        let has_export = self
-            .ctx
-            .arena
-            .has_modifier(modifiers, SyntaxKind::ExportKeyword);
-        let has_default = self
-            .ctx
-            .arena
-            .has_modifier(modifiers, SyntaxKind::DefaultKeyword);
-        if has_export && has_default {
-            return true;
-        }
-        // Check if parent is an export default (ExportDeclaration with is_default_export)
-        if let Some(ext) = self.ctx.arena.get_extended(class_idx)
-            && let Some(parent) = self.ctx.arena.get(ext.parent)
-        {
-            if parent.kind == syntax_kind_ext::EXPORT_DECLARATION
-                && let Some(export_data) = self.ctx.arena.get_export_decl(parent)
-                && export_data.is_default_export
-            {
-                return true;
-            }
-            // Also check for ExportAssignment (export = class { ... })
-            if parent.kind == syntax_kind_ext::EXPORT_ASSIGNMENT {
-                return true;
-            }
-        }
-        false
-    }
-
-    /// Return the node index to anchor TS4094 at for an exported anonymous class.
-    ///
-    /// tsc reports TS4094 at the `export` keyword (col 1), not the `class` keyword.
-    /// When the class is an expression inside an export statement, the parent node
-    /// starts at `export`. When it's a `ClassDeclaration` with own `export default`
-    /// modifiers, the first modifier starts before the class keyword.
-    fn get_anonymous_class_export_anchor(
-        &self,
-        class_idx: NodeIndex,
-        modifiers: &Option<tsz_parser::parser::NodeList>,
-    ) -> Option<NodeIndex> {
-        use tsz_scanner::SyntaxKind;
-        let has_export = self
-            .ctx
-            .arena
-            .has_modifier(modifiers, SyntaxKind::ExportKeyword);
-        let has_default = self
-            .ctx
-            .arena
-            .has_modifier(modifiers, SyntaxKind::DefaultKeyword);
-        if has_export && has_default {
-            // ClassDeclaration with `export default` modifiers. Use the first modifier
-            // node as the anchor so we report at `export` (col 1), not `class`.
-            if let Some(mods) = modifiers
-                && let Some(&first_mod_idx) = mods.nodes.first()
-            {
-                return Some(first_mod_idx);
-            }
-            return Some(class_idx);
-        }
-        // ClassExpression in `export default class` or `export = class`.
-        // The parent export-statement node starts at the `export` keyword.
-        if let Some(ext) = self.ctx.arena.get_extended(class_idx)
-            && let Some(parent) = self.ctx.arena.get(ext.parent)
-        {
-            if parent.kind == syntax_kind_ext::EXPORT_DECLARATION
-                && let Some(export_data) = self.ctx.arena.get_export_decl(parent)
-                && export_data.is_default_export
-            {
-                return Some(ext.parent);
-            }
-            if parent.kind == syntax_kind_ext::EXPORT_ASSIGNMENT {
-                return Some(ext.parent);
-            }
-        }
-        None
-    }
-
-    /// TS4094: Named exported class whose `extends` heritage resolves to an anonymous
-    /// class type.  The anonymous base's private/protected members appear in the .d.ts
-    /// type literal and must be reported.  Errors are anchored at the named class's
-    /// name identifier (matching tsc's anchor position).
-    fn check_ts4094_named_class_anonymous_heritage(
-        &mut self,
-        _class_idx: tsz_parser::parser::NodeIndex,
-        class: &tsz_parser::parser::node::ClassData,
-    ) {
-        let Some(ref heritage_list) = class.heritage_clauses else {
-            return;
-        };
-        for &clause_idx in &heritage_list.nodes {
-            let Some(clause_node) = self.ctx.arena.get(clause_idx) else {
-                continue;
-            };
-            let Some(heritage) = self.ctx.arena.get_heritage_clause(clause_node) else {
-                continue;
-            };
-            // Only `extends` clauses carry the anonymous constructor type.
-            if heritage.token != tsz_scanner::SyntaxKind::ExtendsKeyword as u16 {
-                continue;
-            }
-            for &type_ref_idx in &heritage.types.nodes {
-                let Some(type_ref_node) = self.ctx.arena.get(type_ref_idx) else {
-                    continue;
-                };
-                // Mirror the existing heritage-resolution pattern: if the node is an
-                // ExpressionWithTypeArguments, unpack it; otherwise treat the node itself
-                // as the expression (handles bare identifier heritage like `extends Foo`).
-                let (expr_idx, type_args) =
-                    if let Some(eta) = self.ctx.arena.get_expr_type_args(type_ref_node) {
-                        (eta.expression, eta.type_arguments.as_ref())
-                    } else {
-                        (type_ref_idx, None)
-                    };
-                let Some(base_instance_type) =
-                    self.base_instance_type_from_expression(expr_idx, type_args)
-                else {
-                    continue;
-                };
-                if self.instance_type_is_from_anonymous_class(base_instance_type) {
-                    self.report_instance_type_private_members_as_ts4094(
-                        class.name,
-                        base_instance_type,
-                    );
-                }
-            }
-        }
-    }
-
     /// TS1497: Check that a decorator expression follows the valid grammar.
     ///
     /// Valid decorator expressions are:
@@ -2526,542 +1969,5 @@ impl<'a> CheckerState<'a> {
                 diagnostic_codes::EXPRESSION_MUST_BE_ENCLOSED_IN_PARENTHESES_TO_BE_USED_AS_A_DECORATOR,
             );
         }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::context::CheckerOptions;
-    use crate::test_utils::{check_source, check_source_diagnostics};
-
-    fn check_with_declaration(source: &str) -> Vec<u32> {
-        check_source(
-            source,
-            "test.ts",
-            CheckerOptions {
-                emit_declarations: true,
-                ..CheckerOptions::default()
-            },
-        )
-        .iter()
-        .map(|d| d.code)
-        .collect()
-    }
-
-    // TS4094 — property of exported anonymous class type may not be private or protected
-
-    #[test]
-    fn ts4094_export_default_anon_class_private_member() {
-        // `export default class { private x() {} }` — tsc emits TS4094 for `x`.
-        let codes = check_with_declaration(
-            r#"
-export default class {
-    private x() {}
-    protected y() {}
-    public z() {}
-}
-"#,
-        );
-        assert!(
-            codes.contains(&4094),
-            "TS4094 expected for private/protected in exported anonymous class, got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn ts4094_no_error_for_named_export_own_class() {
-        // `export class Foo { private x() {} }` — tsc does NOT emit TS4094 because
-        // the named class's private members are stripped in the .d.ts.
-        let codes = check_with_declaration(
-            r#"
-export class Foo {
-    private x() {}
-}
-"#,
-        );
-        assert!(
-            !codes.contains(&4094),
-            "TS4094 should NOT fire for named exported class with own private members, got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn ts4094_export_default_mixin_call_anon_class() {
-        // `export default mix(AnonClass)` where mix<T>(x:T):T returns the anonymous
-        // constructor — tsc emits TS4094 for private/protected members.
-        let codes = check_with_declaration(
-            r#"
-declare function mix<TMix>(mixin: TMix): TMix;
-const AnonBase = class {
-    protected _onDispose() {}
-    private _assertIsStripped() {}
-};
-export default mix(AnonBase);
-"#,
-        );
-        assert!(
-            codes.contains(&4094),
-            "TS4094 expected for `export default mix(AnonClass)`, got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn ts4094_named_class_extending_mixin_anon_class() {
-        // `export class Monitor extends mix(AnonBase)` — tsc emits TS4094 at Monitor's
-        // name because the anonymous base's private/protected appear in the .d.ts.
-        let codes = check_with_declaration(
-            r#"
-declare function mix<TMix>(mixin: TMix): TMix;
-const AnonBase = class {
-    protected _onDispose() {}
-    private _assertIsStripped() {}
-};
-export class Monitor extends mix(AnonBase) {
-    protected _onDispose() {}
-}
-"#,
-        );
-        assert!(
-            codes.contains(&4094),
-            "TS4094 expected for named exported class extending mixin of anonymous class, got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn ts4094_exported_function_returning_anon_class_with_private_base() {
-        // The inferred return type of an exported function can be an anonymous
-        // class constructor whose declaration emit surface includes private
-        // members inherited from the base.
-        let codes = check_with_declaration(
-            r#"
-export class Base {
-    private property = "";
-}
-
-export type Constructor<T> = new (...args: any[]) => T;
-export function WithTags<T extends Constructor<Base>>(BaseCtor: T) {
-    return class extends BaseCtor {
-        tags(): void {}
-    };
-}
-
-export class Test extends WithTags(Base) {}
-"#,
-        );
-        let ts4094_count = codes.iter().filter(|&&code| code == 4094).count();
-        assert!(
-            ts4094_count >= 2,
-            "TS4094 expected for exported function return and named class heritage, got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn ts4094_no_error_without_declaration_flag() {
-        // Without `declaration: true`, TS4094 should not be emitted.
-        let codes: Vec<u32> = check_source_diagnostics(
-            r#"
-export default class {
-    private x() {}
-}
-"#,
-        )
-        .iter()
-        .map(|d| d.code)
-        .collect();
-        assert!(
-            !codes.contains(&4094),
-            "TS4094 should NOT fire without declaration flag, got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn ts1267_abstract_property_with_initializer() {
-        let diags = check_source_diagnostics(
-            r#"
-abstract class C {
-    abstract x: number = 1;
-}
-"#,
-        );
-        let matching: Vec<_> = diags.iter().filter(|d| d.code == 1267).collect();
-        assert_eq!(
-            matching.len(),
-            1,
-            "Expected 1 TS1267, got: {:?}",
-            diags.iter().map(|d| d.code).collect::<Vec<_>>()
-        );
-        assert!(matching[0].message_text.contains("'x'"));
-    }
-
-    #[test]
-    fn ts1267_abstract_property_without_initializer_no_error() {
-        let diags = check_source_diagnostics(
-            r#"
-abstract class C {
-    abstract x: number;
-}
-"#,
-        );
-        let matching: Vec<_> = diags.iter().filter(|d| d.code == 1267).collect();
-        assert_eq!(matching.len(), 0, "Expected no TS1267, got: {matching:?}");
-    }
-
-    #[test]
-    fn ts18006_field_named_constructor() {
-        let diags = check_source_diagnostics(
-            r#"
-class C {
-    "constructor" = 3;
-}
-"#,
-        );
-        let matching: Vec<_> = diags.iter().filter(|d| d.code == 18006).collect();
-        assert_eq!(
-            matching.len(),
-            1,
-            "Expected 1 TS18006, got: {:?}",
-            diags.iter().map(|d| d.code).collect::<Vec<_>>()
-        );
-    }
-
-    #[test]
-    fn ts2699_static_property_named_prototype() {
-        let diags = check_source_diagnostics(
-            r#"
-class C {
-    static prototype: number;
-}
-"#,
-        );
-        let matching: Vec<_> = diags.iter().filter(|d| d.code == 2699).collect();
-        assert_eq!(
-            matching.len(),
-            1,
-            "Expected 1 TS2699, got: {:?}",
-            diags.iter().map(|d| d.code).collect::<Vec<_>>()
-        );
-    }
-
-    #[test]
-    fn ts2699_non_static_prototype_no_error() {
-        let diags = check_source_diagnostics(
-            r#"
-class C {
-    prototype: number;
-}
-"#,
-        );
-        let matching: Vec<_> = diags.iter().filter(|d| d.code == 2699).collect();
-        assert_eq!(
-            matching.len(),
-            0,
-            "Expected no TS2699 for non-static prototype, got: {matching:?}"
-        );
-    }
-
-    #[test]
-    fn ts2797_mixin_extending_abstract_type_variable() {
-        let diags = check_source_diagnostics(
-            r#"
-function Mixin<TBaseClass extends abstract new (...args: any) => any>(baseClass: TBaseClass) {
-    class MixinClass extends baseClass {
-    }
-    return MixinClass;
-}
-"#,
-        );
-        let all_codes: Vec<_> = diags.iter().map(|d| d.code).collect();
-        let matching: Vec<_> = diags.iter().filter(|d| d.code == 2797).collect();
-        assert_eq!(
-            matching.len(),
-            1,
-            "Expected 1 TS2797 for mixin extending abstract type variable, got codes: {all_codes:?}"
-        );
-    }
-
-    #[test]
-    fn ts2797_mixin_with_implements_clause() {
-        // TS2797 should still fire when the mixin class has an implements clause
-        // (previously broken due to name-merging between function Mixin and interface Mixin)
-        let diags = check_source_diagnostics(
-            r#"
-interface Mixin {
-    mixinMethod(): void;
-}
-function Mixin<TBaseClass extends abstract new (...args: any) => any>(baseClass: TBaseClass): TBaseClass & (abstract new (...args: any) => Mixin) {
-    class MixinClass extends baseClass implements Mixin {
-        mixinMethod() {}
-    }
-    return MixinClass;
-}
-"#,
-        );
-        let all_codes: Vec<_> = diags.iter().map(|d| d.code).collect();
-        let matching: Vec<_> = diags.iter().filter(|d| d.code == 2797).collect();
-        assert_eq!(
-            matching.len(),
-            1,
-            "Expected 1 TS2797 for mixin with implements clause, got codes: {all_codes:?}"
-        );
-    }
-
-    #[test]
-    fn ts2515_expression_based_heritage() {
-        // Non-abstract class extending expression-based heritage (mixin pattern)
-        // should report TS2515 for unimplemented abstract members
-        let diags = check_source_diagnostics(
-            r#"
-interface Mixin {
-    mixinMethod(): void;
-}
-function Mixin<TBaseClass extends abstract new (...args: any) => any>(baseClass: TBaseClass): TBaseClass & (abstract new (...args: any) => Mixin) {
-    class MixinClass extends baseClass implements Mixin {
-        mixinMethod() {}
-    }
-    return MixinClass;
-}
-
-abstract class AbstractBase {
-    abstract abstractBaseMethod(): void;
-}
-
-const MixedBase = Mixin(AbstractBase);
-
-class DerivedFromAbstract extends MixedBase {
-}
-"#,
-        );
-        let all_codes: Vec<_> = diags.iter().map(|d| d.code).collect();
-        let ts2515: Vec<_> = diags.iter().filter(|d| d.code == 2515).collect();
-        assert_eq!(
-            ts2515.len(),
-            1,
-            "Expected 1 TS2515 for missing abstract member, got codes: {all_codes:?}"
-        );
-        // Verify the message mentions the correct base class name
-        let msg = &ts2515[0].message_text;
-        assert!(
-            msg.contains("AbstractBase & Mixin"),
-            "TS2515 message should reference 'AbstractBase & Mixin', got: {msg}"
-        );
-    }
-
-    /// Rule: a get/set accessor pair (two declarations, one member name) is a
-    /// single inherited abstract member. The missing-member set must dedup by
-    /// name so the count stays 1 (TS2515 singular), not 2 (TS2654 plural with a
-    /// duplicated name).
-    #[test]
-    fn abstract_get_set_pair_counted_as_one_member_ts2515() {
-        let diags = check_source_diagnostics(
-            r#"
-abstract class A { abstract get x(): number; abstract set x(v: number); }
-class B extends A {}
-"#,
-        );
-        let codes: Vec<_> = diags.iter().map(|d| d.code).collect();
-        assert!(
-            codes.contains(&2515),
-            "Abstract get/set pair missing in subclass must emit singular TS2515, got: {codes:?}"
-        );
-        assert!(
-            !codes.contains(&2654),
-            "Abstract get/set pair must not be counted as two members (TS2654), got: {codes:?}"
-        );
-        let msg = &diags
-            .iter()
-            .find(|d| d.code == 2515)
-            .expect("expected a TS2515 diagnostic")
-            .message_text;
-        assert!(
-            !msg.contains("'x', 'x'") && !msg.contains("x, x"),
-            "TS2515 message must name member 'x' once, got: {msg}"
-        );
-    }
-
-    /// Same rule, different identifiers — proves the dedup is structural, not
-    /// keyed to the spelling `x`.
-    #[test]
-    fn abstract_get_set_pair_renamed_counted_as_one_member_ts2515() {
-        let diags = check_source_diagnostics(
-            r#"
-abstract class Shape { abstract get area(): number; abstract set area(v: number); }
-class Square extends Shape {}
-"#,
-        );
-        let codes: Vec<_> = diags.iter().map(|d| d.code).collect();
-        assert!(
-            codes.contains(&2515) && !codes.contains(&2654),
-            "Renamed abstract get/set pair must emit singular TS2515, got: {codes:?}"
-        );
-        let msg = &diags
-            .iter()
-            .find(|d| d.code == 2515)
-            .expect("expected a TS2515 diagnostic")
-            .message_text;
-        assert!(
-            !msg.contains("'area', 'area'") && !msg.contains("area, area"),
-            "TS2515 message must name member 'area' once, got: {msg}"
-        );
-    }
-
-    /// Generalization: abstract method overload signatures share one member
-    /// name across two declarations and likewise collapse to a single missing
-    /// member.
-    #[test]
-    fn abstract_method_overload_signatures_counted_as_one_member_ts2515() {
-        let diags = check_source_diagnostics(
-            r#"
-abstract class A {
-    abstract f(): void;
-    abstract f(x: number): void;
-}
-class B extends A {}
-"#,
-        );
-        let codes: Vec<_> = diags.iter().map(|d| d.code).collect();
-        assert!(
-            codes.contains(&2515) && !codes.contains(&2654),
-            "Abstract method overload signatures must collapse to one missing member (TS2515), got: {codes:?}"
-        );
-    }
-
-    /// tsc accepts implementing only the getter of an abstract get/set pair.
-    #[test]
-    fn abstract_get_set_pair_satisfied_by_getter_only() {
-        let diags = check_source_diagnostics(
-            r#"
-abstract class A { abstract get x(): number; abstract set x(v: number); }
-class B extends A { get x(): number { return 0; } }
-"#,
-        );
-        let codes: Vec<_> = diags.iter().map(|d| d.code).collect();
-        assert!(
-            !codes.contains(&2515) && !codes.contains(&2654),
-            "Implementing the getter of an abstract get/set pair must satisfy it, got: {codes:?}"
-        );
-    }
-
-    /// Negative control: two genuinely distinct abstract members stay TS2654
-    /// listing both names.
-    #[test]
-    fn two_distinct_abstract_members_stay_ts2654() {
-        let diags = check_source_diagnostics(
-            r#"
-abstract class A { abstract a(): void; abstract b(): void; }
-class B extends A {}
-"#,
-        );
-        let codes: Vec<_> = diags.iter().map(|d| d.code).collect();
-        assert!(
-            codes.contains(&2654) && !codes.contains(&2515),
-            "Two distinct missing abstract members must emit TS2654, got: {codes:?}"
-        );
-        let msg = &diags
-            .iter()
-            .find(|d| d.code == 2654)
-            .expect("expected a TS2654 diagnostic")
-            .message_text;
-        assert!(
-            msg.contains("'a'") && msg.contains("'b'"),
-            "TS2654 must list both 'a' and 'b', got: {msg}"
-        );
-    }
-
-    /// Negative control: a single missing abstract method stays TS2515.
-    #[test]
-    fn single_abstract_method_stays_ts2515() {
-        let diags = check_source_diagnostics(
-            r#"
-abstract class A { abstract a(): void; }
-class B extends A {}
-"#,
-        );
-        let codes: Vec<_> = diags.iter().map(|d| d.code).collect();
-        assert!(
-            codes.contains(&2515) && !codes.contains(&2654),
-            "A single missing abstract method must emit TS2515, got: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn double_mixin_conditional_type_base_class_has_no_extra_ts2345() {
-        let diags = check_source_diagnostics(
-            r#"
-type Constructor = new (...args: any[]) => {};
-declare const Object: Constructor;
-
-const Mixin1 = <C extends Constructor>(Base: C) => class extends Base { private _fooPrivate!: {}; };
-
-type FooConstructor = typeof Mixin1 extends (a: Constructor) => infer Cls ? Cls : never;
-const Mixin2 = <C extends FooConstructor>(Base: C) => class extends Base {};
-
-class C extends Mixin2(Mixin1(Object)) {}
-"#,
-        );
-        let codes: Vec<_> = diags.iter().map(|d| d.code).collect();
-        assert!(
-            !codes.contains(&2345),
-            "Expected no TS2345 for double mixin conditional base class, got codes: {codes:?}"
-        );
-    }
-
-    #[test]
-    fn ts2545_mixin_with_optional_rest_parameter() {
-        // TS2545: A mixin class must have a constructor with a single rest
-        // parameter of type 'any[]'. Optional rest parameters are invalid.
-        let diags = check_source_diagnostics(
-            r#"
-type Constructor<T = {}> = new (...args?: any[]) => T;
-
-function Mixin<TBase extends Constructor>(Base: TBase) {
-    return class extends Base {
-    };
-}
-"#,
-        );
-        let all_codes: Vec<_> = diags.iter().map(|d| d.code).collect();
-        let matching: Vec<_> = diags.iter().filter(|d| d.code == 2545).collect();
-        assert_eq!(
-            matching.len(),
-            1,
-            "Expected 1 TS2545 for optional rest param in mixin constructor, got codes: {all_codes:?}"
-        );
-    }
-
-    #[test]
-    fn ts2545_no_error_for_valid_mixin_constructor() {
-        // Valid mixin pattern: `...args: any[]` without optional should NOT emit TS2545.
-        let diags = check_source_diagnostics(
-            r#"
-type Constructor = new (...args: any[]) => {};
-
-function Mixin<TBase extends Constructor>(Base: TBase) {
-    return class extends Base {};
-}
-"#,
-        );
-        let all_codes: Vec<_> = diags.iter().map(|d| d.code).collect();
-        assert!(
-            !all_codes.contains(&2545),
-            "Should NOT emit TS2545 for valid mixin constructor pattern, got codes: {all_codes:?}"
-        );
-    }
-
-    #[test]
-    fn ts2545_no_error_for_bare_any_rest_parameter() {
-        // `...args: any` (bare any, not any[]) is also valid for mixin constructors.
-        let diags = check_source_diagnostics(
-            r#"
-function Mixin<TBase extends abstract new (...args: any) => any>(baseClass: TBase) {
-    abstract class MixinClass extends baseClass {}
-    return MixinClass;
-}
-"#,
-        );
-        let all_codes: Vec<_> = diags.iter().map(|d| d.code).collect();
-        assert!(
-            !all_codes.contains(&2545),
-            "Should NOT emit TS2545 for bare `any` rest param type, got codes: {all_codes:?}"
-        );
     }
 }
